% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_kl.R
\name{.merge_opts}
\alias{.merge_opts}
\title{Merge user options over defaults (named lists)}
\usage{
.merge_opts(defaults, opts = list())
}
\description{
Merge user options over defaults (named lists)
}
\keyword{internal}
