% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admvn-package.R
\docType{package}
\name{admvn-package}
\alias{admvn}
\alias{admvn-package}
\title{Multivariate normal CDF with CppAD derivatives}
\description{
Genz quasi-Monte Carlo MVN probabilities with gradient and Hessian.
}
\author{
\strong{Maintainer}: Patrick Brown \email{patrick.brown@gmail.com} [copyright holder]

}
