% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admvn-package.R
\docType{package}
\name{admvn-package}
\alias{admvn}
\alias{admvn-package}
\title{Multivariate normal CDF and SUN density with CppAD derivatives}
\description{
Genz quasi-Monte Carlo MVN probabilities with gradient and Hessian,
SUN(3,3)/SUN(4,4) log-densities with derivatives with respect to a
block-Cholesky parameter vector, and KL minimisation of SUN approximations
on weighted Gauss-Hermite quadrature.
}
\author{
\strong{Maintainer}: Patrick Brown \email{patrick.brown@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Patrick Brown \email{patrick.brown@gmail.com} [copyright holder]
}

}
