% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sun_params.R
\name{.corr_from_free}
\alias{.corr_from_free}
\title{Correlation matrix from unconstrained hyperspherical free parameters}
\usage{
.corr_from_free(z, d, eps = 0)
}
\description{
Correlation matrix from unconstrained hyperspherical free parameters
}
\keyword{internal}
